multi: Move vspd to internal package.
